refactor: internal_applicationパッケージのバレルファイルを削除
- internal_application.dartバレルファイルを削除
- src/operational_settingsをlib/operational_settingsに移動
- src/operational_settingsをtest/operational_settingsに移動
- 不要なinternal_application_base.dartとinternal_application_test.dartを削除
- 依存ファイルのimport文を更新
