+++ apps/app/lib/composition_root/use_cases/check_maintenance_mode_use_case.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_app/composition_root/repositories/operational_settings_repository.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:internal_application/internal_application.dart';
+import 'package:internal_application/operational_settings/check_maintenance_mode_use_case.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'check_maintenance_mode_use_case.g.dart';
