--- packages/application/lib/internal_application.dart
@@ -1,10 +1,0 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
-library;
-
-export 'src/internal_application_base.dart';
-export 'src/operational_settings/check_force_update_use_case.dart';
-export 'src/operational_settings/check_maintenance_mode_use_case.dart';
-
-// TODO: Export any libraries intended for clients of this package.
