--- packages/application/test/internal_application_test.dart
@@ -1,16 +1,0 @@
-import 'package:internal_application/internal_application.dart';
-import 'package:test/test.dart';
-
-void main() {
-  group('A group of tests', () {
-    final awesome = Awesome();
-
-    setUp(() {
-      // Additional setup goes here.
-    });
-
-    test('First Test', () {
-      expect(awesome.isAwesome, isTrue);
-    });
-  });
-}
